1211 - 댓글 조회/작성/수정/삭제 기능 구현 완료

+++ src/main/java/com/inhatc/dev_folio/project/service/CommentService.java
@@ -23,46 +23,61 @@ public class CommentService {
     private final ProjectRepository projectRepository;
     private final MemberRepository memberRepository;
 
-    public List<CommentDto.View> getComment(Long id) {
-        List<Comment> comments = commentRepository.findByProjectId(id);
+    public List<CommentDto.View> getComment(Long projectId, String email) {
+        List<Comment> comments = commentRepository.findByProjectId(projectId);
 
-        // TODO: 로그인 기능 구현 후 현재 로그인 된 사용자로 교체
-        Member member = memberRepository.findById(2L).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.MEMBER_ID_NOT_FOUND.getMessage()));
-
-        CommentMapper commentMapper = CommentMapper.INSTANCE;
-        List<CommentDto.View> views = new ArrayList<>();
-        for (Comment comment : comments) {
-            CommentDto.View view = commentMapper.commentToView(comment);
-            view.setSelf(comment.getMember().equals(member));
-            views.add(view);
+        // 사용자가 로그인상태가 아닐 때
+        if (email == null) {
+            return CommentMapper.INSTANCE.commentListToViewList(comments);
         }
 
-        return views;
+        // 로그인 상태일 때
+        else {
+            Member member = memberRepository.findByEmail(email).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.MEMBER_EMAIL_NOT_FOUND.getMessage()));
+            List<CommentDto.View> views = new ArrayList<>();
+
+            CommentMapper commentMapper = CommentMapper.INSTANCE;
+            for (Comment comment : comments) {
+                CommentDto.View view = commentMapper.commentToView(comment);
+                view.setSelf(comment.getMember().equals(member));
+                views.add(view);
+            }
+            return views;
+        }
     }
 
-    public void writeComment(Long projectId, CommentDto.Contents contents) {
+    public void writeComment(Long projectId, CommentDto.Contents contents, String email) {
         Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.PROJECT_NOT_FOUND.getMessage()));
+        Member member = memberRepository.findByEmail(email).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.MEMBER_EMAIL_NOT_FOUND.getMessage()));
         Comment newComment = Comment.builder()
                 .contents(contents.getContents())
-                // TODO: 로그인 기능이 완료되면 Member도 집어넣기
-                // .member()
+                .member(member)
                 .project(project)
                 .build();
         commentRepository.save(newComment);
     }
 
-    public void updateComment(Long commentId, CommentDto.Contents contents) {
+    public void updateComment(Long commentId, CommentDto.Contents contents, String email) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.COMMENT_NOT_FOUND.getMessage()));
+        // 댓글이 삭제됐는지 확인
         if (comment.isDeleted()) {
             throw new IllegalStateException(ErrorMessage.COMMENT_ALREADY_DELETED.getMessage());
         }
-        // TODO 로그인된 사용자 판별해서 수정 권한 있는지 확인하기
+        // 권한 있는지 확인
+        Member member = memberRepository.findByEmail(email).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.MEMBER_EMAIL_NOT_FOUND.getMessage()));
+        if (comment.getMember() != member) {
+            throw new IllegalStateException(ErrorMessage.COMMENT_ACCESS_DENIED.getMessage());
+        }
         comment.updateContents(contents.getContents());
         commentRepository.save(comment);
     }
 
-    public void deleteComment(Long commentId) {
-        // TODO 로그인된 사용자 판별해서 삭제 권한 있는지 확인하기
+    public void deleteComment(Long commentId, String email) {
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new EntityNotFoundException(ErrorMessage.COMMENT_NOT_FOUND.getMessage()));
+        // 작성자인지 확인
+        if (!comment.getMember().getEmail().equals(email)) {
+            throw new IllegalStateException(ErrorMessage.COMMENT_ACCESS_DENIED.getMessage());
+        }
         commentRepository.deleteById(commentId);
     }
 }
